fixed hero section and spacing
