+++ src/components/QRCode/QRCode.css
@@ -1,6 +1,11 @@
-.qr-code {
-    padding: 50px;
-    margin: 500px;
-    color: var(--primary);
-    background-color: var(--primary);
+/* .hero-container {
+    align-items: flex;
 }
+
+.qr-code-section {
+    
+}
+
+.hero-text-section {
+
+} */
