+++ src/pages/WinWallPage.css
@@ -29,6 +29,6 @@ body {
 }
 
 textarea {
-    font-family: "Comic Sans MS", "Comic Sans";
+    font-family: var(--main-font);
     font-size: medium;
 }
